mode push
tanya noob

+++ api/handler/fundraiserHandler.js
@@ -1,6 +1,7 @@
 
 const jwt = require('jsonwebtoken')
 const Fundraiser = require('../../Models/Fundraiser');
+const User = require('../../Models/User');
 const { validationResult } = require('express-validator');
 
 //POST a fundraiser
@@ -11,6 +12,7 @@ exports.postFundraiser = async (req, res, next) => {
     const habitat= req.body.habitat;
     const description = req.body.description;
     const status= req.body.status;
+    let creator;
     const fundraiser = new Fundraiser({
         name: name,
         scfname: scfname,
@@ -21,8 +23,21 @@ exports.postFundraiser = async (req, res, next) => {
     });
     return fundraiser
     .save()
+    .then(result=>{
+      return User.findById(req.userId);
+    })
+ 
+  .then(user => {
+    creator = user;
+    user.fundraisers.push(fundraiser);
+    return user.save();
+  })
     .then(result => {
-        res.status(201).json({ message: 'New Fundraiser Creater', FundraiserId: result._id , fundraiser: result});
+        res.status(201).json({ message: 'New Fundraiser Creater', 
+        FundraiserId: fundraiser._id , 
+        fundraiser: fundraiser,
+        creator: { _id: creator._id, name: creator.name }
+      });
       })
       .catch(err => {
         if (!err.statusCode) {
@@ -60,6 +75,7 @@ exports.updateFundraiser = async (req, res, next) => {
   const habitat= req.body.habitat;
   const description = req.body.description;
   const status= req.body.status;
+
   Fundraiser.findById(fundId)
   .then(fundraiser=>{
     if(!fundraiser){
@@ -71,12 +87,17 @@ exports.updateFundraiser = async (req, res, next) => {
     fundraiser.scfname= scfname;
     fundraiser.habitat= habitat;
     fundraiser.description = description;
-    fundraiser.status= status;
-    fundraiser.save()
-
-  })
+    fundraiser.status= status; 
+  });
+    fundraiser
+    .save()
+   
   .then(result => {
-    res.status(200).json({ message: 'Fundraiser updated!', post: result });
+    res.status(200).json({
+      message: 'Fundraiser updated!', 
+      post: result 
+
+    });
   })
   .catch(err => {
     if (!err.statusCode) {
@@ -107,8 +128,28 @@ exports.getOneFundraiser = async (req, res, next) => {
   next(err);
 });
 };
+//Get my fundRaiser
+exports.getMyFundraiser = async (req, res, next) => {
 
-//delet fundraiser
+  const userId= req.params.userId;
+  Fundraiser.findById(fundId)
+  .then(fundraiser=>{
+    if(!fundraiser){
+      const error = new Error('Could not find fundraiser.');
+      error.statusCode = 404;
+      throw error;
+    }
+    res.status(201).json({ message: 'Fundraiser Got', fundraiser: fundraiser});  
+})
+.catch(err => {
+  if (!err.statusCode) {
+    err.statusCode = 500;
+  }
+  next(err);
+});
+}
+
+//delete fundraiser
 exports.deleteFundraiser = (req, res, next) => {
   const fundId= req.params.fundId;
   Fundraiser.findById(fundId)
@@ -118,8 +159,19 @@ exports.deleteFundraiser = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
-
+      if (fundraiser.userId.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
       return Fundraiser.findByIdAndRemove(fundId);
+    })
+    .then(result => {
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      user.fundraisers.pull(fundId);
+      return user.save();
     })
     .then(result => {
       res.status(200).json({ message: 'Deleted fundraiser.' });
